+++ fusion_engine_driver/package.xml
@@ -33,10 +33,6 @@
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
 
-  <exec_depend>rosidl_default_runtime</exec_depend>
-  <member_of_group>rosidl_interface_packages</member_of_group>
-  <build_export_depend>rosidl_default_runtime</build_export_depend>
-
   <export>
     <build_type>ament_cmake</build_type>
   </export>
